Renamed form to MainWindow
