+++ WinFormsStudentScheduler/Forms/MainWindow.cs
@@ -8,11 +8,11 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WinFormsStudentScheduler
+namespace WinFormsStudentScheduler.Forms
 {
-    public partial class MainForm : Form
+    public partial class MainWindow : Form
     {
-        public MainForm()
+        public MainWindow()
         {
             InitializeComponent();
         }
