+++ WinFormsStudentScheduler/Program.cs
@@ -16,7 +16,7 @@ namespace WinFormsStudentScheduler
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new MainForm());
+            Application.Run(new Form1());
         }
     }
 }
